taste_buddy_server/pyproject.toml
[tool.poetry]
name = "taste_buddy_server"
version = "0.1.0"
description = ""
authors = ["am9zZWY <46693545+am9zZWY@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "^3.10.5"
black = "^24.8.0"
cachetools = "^5.5.0"
fastapi = "^0.115.0"
httptools = "^0.6.1"
ingredient-parser-nlp = "^1.1.1"
openai = "^1.46.1"
pip-chill = "^1.0.3"
pydantic = "^2.9.2"
pymongo = "^4.9.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
recipe-scrapers = "^15.1.0"
starlette = "^0.38.5"
uvloop = "^0.20.0"
uvicorn = "^0.30.6"
pint = "^0.24.3"
python-multipart = "^0.0.9"
argon2-cffi = "^23.1.0"
pyjwt = "^2.9.0"
rapidfuzz = "^3.9.7"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.115.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
